+++ src/ska_dlm_client/openapi/models/validation_error.py
@@ -20,26 +20,22 @@ import re  # noqa: F401
 from typing import Any, ClassVar, Dict, List, Optional, Set
 
 from pydantic import BaseModel, ConfigDict, StrictStr
-from pydantic_core import to_jsonable_python
 from typing_extensions import Self
 
-from ska_dlm_client.openapi.models.location_inner import LocationInner
+from ska_dlm_client.openapi.models.validation_error_loc_inner import ValidationErrorLocInner
 
 
 class ValidationError(BaseModel):
     """ValidationError"""
 
     # noqa: E501
-    loc: List[LocationInner]
+    loc: List[ValidationErrorLocInner]
     msg: StrictStr
     type: StrictStr
-    input: Optional[Any] = None
-    ctx: Optional[Dict[str, Any]] = None
-    __properties: ClassVar[List[str]] = ["loc", "msg", "type", "input", "ctx"]
+    __properties: ClassVar[List[str]] = ["loc", "msg", "type"]
 
     model_config = ConfigDict(
-        validate_by_name=True,
-        validate_by_alias=True,
+        populate_by_name=True,
         validate_assignment=True,
         protected_namespaces=(),
     )
@@ -50,7 +46,8 @@ class ValidationError(BaseModel):
 
     def to_json(self) -> str:
         """Returns the JSON representation of the model using alias"""
-        return json.dumps(to_jsonable_python(self.to_dict()))
+        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
+        return json.dumps(self.to_dict())
 
     @classmethod
     def from_json(cls, json_str: str) -> Optional[Self]:
@@ -81,11 +78,6 @@ class ValidationError(BaseModel):
                 if _item_loc:
                     _items.append(_item_loc.to_dict())
             _dict["loc"] = _items
-        # set to None if input (nullable) is None
-        # and model_fields_set contains the field
-        if self.input is None and "input" in self.model_fields_set:
-            _dict["input"] = None
-
         return _dict
 
     @classmethod
@@ -100,14 +92,12 @@ class ValidationError(BaseModel):
         _obj = cls.model_validate(
             {
                 "loc": (
-                    [LocationInner.from_dict(_item) for _item in obj["loc"]]
+                    [ValidationErrorLocInner.from_dict(_item) for _item in obj["loc"]]
                     if obj.get("loc") is not None
                     else None
                 ),
                 "msg": obj.get("msg"),
                 "type": obj.get("type"),
-                "input": obj.get("input"),
-                "ctx": obj.get("ctx"),
             }
         )
         return _obj