+++ tests/entrypoint.sh
@@ -11,6 +11,7 @@ case "$watcher_mode" in
         CMD="dlm-directory-watcher \
           --location ${LOCATION_NAME:-SKA-DEV} \
           --source-name ${SOURCE_NAME:-dir-watcher} \
+          --source-phase ${SOURCE_PHASE:-GAS} \
           --directory-to-watch ${DIRECTORY_TO_WATCH:-/dlm/watch_dir} \
           --target-name ${TARGET_NAME:-dlm-archive} \
           --migration-url ${MIGRATION_URL:-http://dlm_migration:8004} \
@@ -27,6 +28,7 @@ case "$watcher_mode" in
         CMD="dlm-configdb-watcher \
           --location ${LOCATION_NAME:-SKA-DEV} \
           --source-name ${SOURCE_NAME:-configdb-watcher} \
+          --source-phase ${SOURCE_PHASE:-GAS} \
           --directory-to-watch ${DIRECTORY_TO_WATCH:-/dlm/product_dir} \
           --target-name ${TARGET_NAME:-dlm-archive} \
           --storage-url ${STORAGE_URL:-http://dlm_storage:8003} \