+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Updated
 
+* Register-only functionality when SOURCE and TARGET are the same.
 * Get files sizes and pass them to `register_data_item`
 * The default SERVER_IMAGE for testing is now a local dirty build to enable direct startup of containers from the compose file.
 * The DLM heuristics removes test files (no direct `rm` commands)